Test: complete client-backed coverage of the diagnostic publish boundary
Add the two client-backed behavioural assertions the `prepare_publish`-only
tests could not observe, without a mock that bypasses `ClientSocket::notify`:

- **Missing-feature-index skip** (`publish.rs`): an in-process `async_lsp`
  main loop supplies a real `ClientSocket`; `publish_feature_diagnostics` is
  called for a path with no feature index, and the captured outgoing transport
  contains a `window/logMessage` sentinel but no `publishDiagnostics` — proving
  the absent index skips the whole boundary rather than clearing prior client
  diagnostics. Fails if `publish_with` notifies despite the missing index.
- **Rust empty-vector clearing** (`smoke_lsp`): an unused Rust step yields a
  non-empty `publishDiagnostics`; once a feature references it, re-saving the
  Rust file re-publishes an empty array for the same Rust URI. Both the method
  name and URI are asserted so feature notifications cannot satisfy it.

Split the two clearing smoke tests into `tests/smoke_lsp/clearing.rs` to keep
`main.rs` under the 400-line limit, add a `tokio` `io-util` dev-dependency for
the in-process transport, and update the developers' guide to identify the new
transport-backed assertions.

Validated: make check-fmt (pass), make markdownlint (pass), make lint (pass —
clippy + Whitaker + python + file-length), make test (pass — 1511 nextest, 56
pytest).
