+++ crates/rstest-bdd-server/Cargo.toml
@@ -49,3 +49,6 @@ proptest.workspace = true
 rstest.workspace = true
 rstest-bdd-server = { path = ".", features = ["test-support"] }
 tempfile.workspace = true
+# `io-util` supplies `duplex`/`empty`/`AsyncReadExt` for the in-process
+# transport-backed diagnostic-publishing test.
+tokio = { workspace = true, features = ["io-util"] }
